+++ src/pages/AudioBooks.js
@@ -1,0 +1,162 @@
+import React from "react";
+
+function AudioBooks() {
+  return (
+    <div>
+      <h1>Disclaimer </h1>
+
+      <ol>
+        By navigating away from this website, you acknowledge:
+        <br />
+        <li>
+          No Endorsement or Responsibility: We are not responsible for the content, accuracy, or legality of external websites. Your decision to leave
+          is at your own risk.{" "}
+        </li>
+        <li>
+          Third-Party Websites: Links to third-party sites are for convenience only. We have no control over their content and make no guarantees
+          about their accuracy or security.
+        </li>
+        <li>
+          No Warranty or Guarantee: We make no warranties about the completeness, accuracy, or reliability of the information on this website. Your
+          reliance on such information is at your own risk.
+        </li>
+        <li> Release from Liability: By leaving, you release us from any liability related to your use of external websites.</li>
+        <li> Changes to Disclaimer: This disclaimer is subject to change without notice.</li>
+      </ol>
+
+      <h2>Audio Books on YouTube</h2>
+      <ol>
+        <li>
+          <a href='https://www.youtube.com/watch?v=2QLy4_2spkk' target='_blank' rel='noreferrer'>
+            1984
+          </a>{" "}
+          by George Orwell
+        </li>
+
+        <li>
+          <a
+            href='https://www.youtube.com/results?search_query=%22All+Quiet+on+the+Western+Front%22+by+Erich+Maria+Remarque+audiobook'
+            target='_blank'
+            rel='noreferrer'>
+            All Quiet on the Western Front
+          </a>{" "}
+          by Erich Maria Remarque
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=bG5QixsoPok' target='_blank' rel='noreferrer'>
+            Fahrenheit 451
+          </a>{" "}
+          by Ray Bradbury
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=tETWQO2UXmQ' target='_blank' rel='noreferrer'>
+            Frankenstein
+          </a>{" "}
+          by Mary Shelley
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=IsZsIjMz5Cg' target='_blank' rel='noreferrer'>
+            Harry Potter and the Sorcerer's Stone
+          </a>{" "}
+          by J.K. Rowling
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=GwsOgHASmyU' target='_blank' rel='noreferrer'>
+            Of Mice and Men
+          </a>{" "}
+          by John Steinbeck
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=utRA2Ck3swU' target='_blank' rel='noreferrer'>
+            Lord of the Flies
+          </a>{" "}
+          by William Golding
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=e4tSfkGyC74' target='_blank' rel='noreferrer'>
+            Romeo And Juliet
+          </a>{" "}
+          by William Shakespeare
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=Yz6uGkdenEk' target='_blank' rel='noreferrer'>
+            Slaughterhouse Five
+          </a>{" "}
+          by Kurt Vonnegut
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=VKNGnXA1ix8' target='_blank' rel='noreferrer'>
+            The Adventures of Huckleberry Finn
+          </a>{" "}
+          by Mark Twain
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=ot74dEJKADg' target='_blank' rel='noreferrer'>
+            The Catcher in the Rye
+          </a>{" "}
+          by J.D. Salinger
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=vVL9TtxDu60' target='_blank' rel='noreferrer'>
+            The Color Purple
+          </a>{" "}
+          by Alice Walker
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=ZjHl5TyWMsQ' target='_blank' rel='noreferrer'>
+            The Great Gatsby
+          </a>{" "}
+          by F.Scott Fitzgerald
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=0pmWf1k-vFM' target='_blank' rel='noreferrer'>
+            The Handmaid's Tale
+          </a>{" "}
+          by Margaret Atwood
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=xtoS738ckCI&list=PLCiJVrVHvzoLzV1FQlvlduj2kGmd75mC0' target='_blank' rel='noreferrer'>
+            The Hobbit
+          </a>{" "}
+          by J.R.R. Tolkien
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=JXxy63IW7-Y' target='_blank' rel='noreferrer'>
+            The Old Man And The Sea
+          </a>{" "}
+          by Ernest Hemingway
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=udvHrcJGIAA' target='_blank' rel='noreferrer'>
+            The Scarlet Letter
+          </a>{" "}
+          by Nathaniel Hawthorne
+        </li>
+
+        <li>
+          <a href='https://www.youtube.com/watch?v=XODFEs8Gspk' target='_blank' rel='noreferrer'>
+            To Kill A Mockingbird
+          </a>{" "}
+          by Harper Lee
+        </li>
+      </ol>
+    </div>
+  );
+}
+
+export default AudioBooks;